Extracting Entity into a trait, the code is reusable for any other type or datasource
